requirements.txt
google-genai
PyGithub
python-dotenv
pydantic
